+++ .gitignore
@@ -9,4 +9,6 @@
 *.idea
 .molecule
 .cache
-**__pycache__**
+__pycache__/
+.pytest_cache
+.tox
