+++ .travis.yml
@@ -1,18 +1,25 @@
+---
 sudo: required
 language: python
 cache: pip
 services:
   - docker
 env:
-  - ANSIBLE='ansible>=2.2.0,<2.3.0'
-  - ANSIBLE='ansible>=2.3.0,<2.4.0'
-  - ANSIBLE='ansible>=2.4.0,<2.5.0'
+  - ANSIBLE=2.4
+  - ANSIBLE=2.5
+  - ANSIBLE=2.6
+matrix:
+  fast_finish: true
 install:
-  - pip install ${ANSIBLE} 'ansible-lint>=3.4.15' 'molecule==1.25.0' docker git-semver 'testinfra>=1.7.0,<=1.10.1'
-before_script:
-  - ansible-lint tests/playbook.yml
+  - pip install tox-travis git-semver
 script:
-  - molecule test
+  - tox
+deploy:
+  provider: script
+  skip_cleanup: true
+  script: .travis/releaser.sh
+  on:
+    branch: master
 branches:
   only:
     - master
