+++ meta/main.yml
@@ -1,16 +1,12 @@
 galaxy_info:
   author: Geoffrey Bachelot
   description: Install netdata on Linux host
-  company: HSFactory
-
-  license: license Apache v2.0
-
-  min_ansible_version: 2.2
-
+  license: BSD
+  min_ansible_version: 2.4
+  role_name: netdata
   platforms:
     - name: EL
       versions:
-        - 6
         - 7
     - name: Ubuntu
       versions:
@@ -20,17 +16,11 @@ galaxy_info:
       versions:
         - jessie
         - wheezy
-
-  galaxy_tags: [netdata]
-    # List tags for your role here, one per line. A tag is
-    # a keyword that describes and categorizes the role.
-    # Users find roles by searching for tags. Be sure to
-    # remove the '[]' above if you add tags to this list.
-    #
-    # NOTE: A tag is limited to a single word comprised of
-    # alphanumeric characters. Maximum 20 tags per role.
+    - name: Fedora
+      versions:
+        - 27
+  galaxy_tags:
+    - netdata
+    - monitoring
 
 dependencies: []
-  # List your role dependencies here, one per line.
-  # Be sure to remove the '[]' above if you add dependencies
-  # to this list.
