+++ molecule/default/playbook.yml
@@ -1,5 +1,5 @@
 ---
 - hosts: all
-  any_errors_fatal: yes
+  any_errors_fatal: true
   roles:
     - ansible-netdata
